Move `library` folder
