+++ client/web/src/enterprise/batches/batch-spec/edit/library/hello-world.batch.yaml
