+++ client/web/src/enterprise/batches/batch-spec/edit/library/minimal.batch.yaml
